Tidy and move files.
