+++ packages/markuplint/src/cli/bootstrap.ts
